+++ netcfgbu/consts.py
@@ -6,7 +6,7 @@ DEFAULT_PROBE_TIMEOUT = 10
 # DEFAULT_CONFIG_STARTS_AFTER = "Current configuration"
 # DEFAULT_CONFIG_ENDS_WITH = "end"
 
-PROMPT_VALID_CHARS = r"a-z0-9.\-_@()/:"
+PROMPT_VALID_CHARS = r"a-z0-9.\-_@()/:~"
 PROMPT_MAX_CHARS = 65
 
 INVENTORY_FIELDNAMES = ["host", "ipaddr", "os_name", "username", "password"]